splitting of "src/core" modules into smaller sub modules
