+++ src/core/clustering/__init__.py
@@ -1,0 +1,10 @@
+"""Clustering and concept-graph clustering helpers."""
+
+from src.core.clustering.phrase import ClusterNumberDetector, PhraseClusterFactory
+from src.core.clustering.word_embedding import WordEmbeddingClustering
+
+__all__ = [
+    "ClusterNumberDetector",
+    "PhraseClusterFactory",
+    "WordEmbeddingClustering",
+]
