+++ src/core/embedding_functions.py
@@ -10,7 +10,7 @@ from sklearn.cluster import KMeans
 
 from src.common.io import load_pickle, save_pickle
 from src.common.spacy_extensions import set_spacy_extensions
-from src.core.data_functions import DataProcessingFactory
+from src.core.data import DataProcessingFactory
 from src.core.reduction import NoneDownScaleObj
 from src.storage.marqo_external_utils import MarqoEmbeddingStore
