+++ src/pipeline/load_utils.py
@@ -5,8 +5,8 @@ from typing import Optional, Union
 import networkx as nx
 
 from src.common.io import load_pickle
-from src.core.cluster_functions import PhraseClusterFactory
-from src.core.data_functions import DataProcessingFactory
+from src.core.clustering import PhraseClusterFactory
+from src.core.data import DataProcessingFactory
 from src.core.embedding_functions import SentenceEmbeddingsFactory
 from src.pipeline.status import StepsName
